+++ admin/gererEvenements.php
@@ -992,7 +992,7 @@ echo $verif->getErreur("genre");
 echo "<option value=\"0\">&nbsp;</option>";
 $req_lieux = $connector->query("
 SELECT idLieu, nom FROM lieu
-WHERE actif=1 AND statut='actif'
+WHERE statut='actif'
 ORDER BY TRIM(LEADING 'L\'' FROM (TRIM(LEADING 'Les ' FROM (TRIM(LEADING 'La ' FROM (TRIM(LEADING 'Le ' FROM nom)))))))
 COLLATE utf8mb4_unicode_ci"
  );